change double to float

+++ includes/color.h
@@ -6,7 +6,7 @@
 /*   By: llai <llai@student.42london.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/29 15:22:50 by llai              #+#    #+#             */
-/*   Updated: 2024/05/02 15:56:10 by llai             ###   ########.fr       */
+/*   Updated: 2024/05/03 17:11:16 by llai             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,10 +15,10 @@
 
 typedef struct s_color
 {
-	double	transparent;
-	double	red;
-	double	green;
-	double	blue;
+	float	transparent;
+	float	red;
+	float	green;
+	float	blue;
 }	t_color;
 
 int		create_trgb(int t, int r, int g, int b);
@@ -26,11 +26,11 @@ int		get_t(int trgb);
 int		get_r(int trgb);
 int		get_g(int trgb);
 int		get_b(int trgb);
-t_color	color(double transparent, double red, double green, double blue);
-int		create_trgb_color(double t, double r, double g, double b);
+t_color	color(float transparent, float red, float green, float blue);
+int		create_trgb_color(float t, float r, float g, float b);
 t_color	add_colors(t_color c1, t_color c2);
 t_color	sub_colors(t_color c1, t_color c2);
-t_color	mul_color(t_color c, double scalar);
+t_color	mul_color(t_color c, float scalar);
 t_color	hadamard_product(t_color c1, t_color c2);
 void	print_color(t_color color);
